+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Code based features to add]
 1. Refactor this code to shorter with `.Query().Get()` method, and the error with `http.NotFound()` method.
+2. Create `type.go` in `database/users` repo.
 
 ## [App functionalities to add]
 1. User ID details.